refactor get deps and fix tests
