+++ epochalyst/pipeline/model/training/utils/__init__.py
@@ -1,1 +1,8 @@
 """Module with utility functions for training."""
+
+from .get_dependencies import _get_onnxrt, _get_openvino
+
+__all__ = [
+    "_get_onnxrt",
+    "_get_openvino",
+]
