+++ epochalyst/pipeline/model/training/utils/get_dependencies.py
@@ -1,0 +1,31 @@
+"""Defines function for importing onnxruntime."""
+
+from typing import Any
+
+
+def _get_onnxrt() -> Any:  # noqa: ANN401
+    """Return onnxruntime."""
+    try:
+        import onnxruntime as onnxrt
+
+    except ImportError:
+        raise ImportError(
+            "If you want to compile models to ONNX format install onnxruntime",
+        ) from None
+
+    else:
+        return onnxrt
+
+
+def _get_openvino() -> Any:  # noqa: ANN401
+    """Return openvino."""
+    try:
+        import openvino
+
+    except ImportError:
+        raise ImportError(
+            "If you want to compile models to OpenVINO format install openvino",
+        ) from None
+
+    else:
+        return openvino
