--- epochalyst/pipeline/model/training/utils/get_onnxrt.py
@@ -1,16 +1,0 @@
-"""Defines function for importing onnxruntime."""
-from typing import Any
-
-
-def get_onnxrt() -> Any:  # noqa: ANN401
-    """Return onnxruntime."""
-    try:
-        import onnxruntime as onnxrt
-
-    except ImportError:
-        raise ImportError(
-            "If you want to compile models to ONNX format install onnxruntime",
-        ) from None
-
-    else:
-        return onnxrt
